daemon: sortMounts: move to container.SortMounts
Move the utility to the container package, which is where the Mount
type is defined, and to allow re-use of the utility both inside the
container package and the daemon package.
